% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ode_em}
\alias{ode_em}
\title{Euler-Maruyama method (EM)}
\usage{
ode_em(tau = 0.01, noise_strength = 2)
}
\arguments{
\item{tau}{tau parameter}

\item{noise_strength}{noise_strength parameter}
}
\description{
Euler-Maruyama method implementation of the \acronym{ODE}.
}
