% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_reactions.R
\name{port_reactions}
\alias{port_reactions}
\title{Port GillespieSSA parameters to GillespieSSA2}
\usage{
port_reactions(a, nu)
}
\arguments{
\item{a}{The \code{a} parameter of \code{\link[GillespieSSA:ssa]{GillespieSSA::ssa()}}.}

\item{nu}{The \code{nu} parameter of \code{\link[GillespieSSA:ssa]{GillespieSSA::ssa()}}.}
}
\description{
This is a helper function to tranform GillesieSSA-style
paramters to GillespieSSA2.
}
\examples{
a   <- c("c1*Y1","c2*Y1*Y2","c3*Y2")
nu  <- matrix(c(+1,-1,0,0,+1,-1),nrow=2,byrow=TRUE)
port_reactions(a, nu)
}
