% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{print.SSA_reaction}
\alias{print.SSA_reaction}
\title{Print a reaction}
\usage{
\method{print}{SSA_reaction}(x, ...)
}
\arguments{
\item{x}{A reaction}

\item{...}{Not used}
}
\description{
Print a reaction
}
\examples{
r <- reaction(~ c1 * s1, c(s1 = -1, s2 = +1))
print(r)
}
