% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa.direct}
\alias{ssa.direct}
\title{Direct method (D)}
\usage{
ssa.direct()
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Direct method implementation of the \acronym{SSA} as described by Gillespie (1977).
}
\references{
Gillespie (1977)
}
\seealso{
\link{fastgssa-package}, \code{\link{ssa}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
