% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa.em}
\alias{ssa.em}
\title{Euler–Maruyama method (EM)}
\usage{
ssa.em(h = 0.01, noise_strength = 2)
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Euler–Maruyama method implementation of https://en.wikipedia.org/wiki/Euler%E2%80%93Maruyama_method.
}
\seealso{
\link{fastgssa-package}, \code{\link{ssa}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}

