% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa.otl}
\alias{ssa.otl}
\title{Optimized tau-leap method (OTL)}
\usage{
ssa.otl(x0, hor = NaN, nc = 10, epsilon = 0.03, dtf = 10, nd = 100)
}
\arguments{
\item{x0}{x0 param}

\item{hor}{highest order reaction vector (one entry per species in \code{x})}

\item{nc}{number of critical reactions threshold parameter.}

\item{epsilon}{error control parameter.}

\item{dtf}{Direct method threshold factor for temporarily suspending the method.}

\item{nd}{number of Direct method steps to perform during a suspension.}
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Optimized tau-leap method implementation of the \acronym{SSA} as described by Cao et al. (2006).
}
\references{
Cao et al. (2006)
}
\seealso{
\link{fastgssa-package}, \code{\link{ssa}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
