% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa_btl}
\alias{ssa_btl}
\title{Binomial tau-leap method (BTL)}
\usage{
ssa_btl(f = 10)
}
\arguments{
\item{f}{coarse-graining factor (see page 4 in Chatterjee et al. 2005).}
}
\value{
an object of to be used by \code{\link[=ssa]{ssa()}}.
}
\description{
Binomial tau-leap method implementation of the \acronym{SSA} as described by Chatterjee et al. (2005).
}
\references{
Chatterjee A., Vlachos D.G., and Katsoulakis M.A. 2005. Binomial distribution based tau-leap accelerated stochastic simulation. J. Chem. Phys. 122:024112. \url{http://dx.doi.org/10.1063/1.1833357}
}
