% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa_direct}
\alias{ssa_direct}
\title{Direct method (D)}
\usage{
ssa_direct()
}
\value{
an object of to be used by \code{\link[=ssa]{ssa()}}.
}
\description{
Direct method implementation of the \acronym{SSA} as described by Gillespie (1977).
}
\references{
Gillespie D.T. 1977. Exact stochastic simulation of coupled chemical reactions. J. Phys. Chem. 81:2340. \url{http://dx.doi.org/10.1021/j100540a008}
}
