% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa_em}
\alias{ssa_em}
\title{Euler-Maruyama method (EM)}
\usage{
ssa_em(tau = 0.01, noise_strength = 2)
}
\arguments{
\item{tau}{tau parameter}

\item{noise_strength}{noise_strength parameter}
}
\description{
Euler-Maruyama method implementation
}
