% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa_etl}
\alias{ssa_etl}
\title{Explicit tau-leap method (ETL)}
\usage{
ssa_etl(tau)
}
\arguments{
\item{tau}{the step-size (default 0.3).}
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Explicit tau-leap method implementation of the \acronym{SSA} as described by Gillespie (2001).
}
\references{
Gillespie (2001)
}
