% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_exact.R
\name{ssa_exact}
\alias{ssa_exact}
\title{Exact method}
\usage{
ssa_exact()
}
\value{
an object of to be used by \code{\link[=ssa]{ssa()}}.
}
\description{
Exact method implementation of the \acronym{SSA} as described by Gillespie (1977).
}
\references{
Gillespie (1977) \url{doi:10.1021/j100540a008}
}
