% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_plot.R
\name{ssa_plot}
\alias{ssa_plot}
\title{Simple plotting of ssa output}
\usage{
ssa_plot(ssa_out)
}
\arguments{
\item{ssa_out}{Data object returned by \code{\link[=ssa]{ssa()}}.}
}
\description{
Provides basic functionally for simple and quick time series plot of simulation output from \code{\link[=ssa]{ssa()}}.
}
