% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa.btl}
\alias{ssa.btl}
\title{Binomial tau-leap method (BTL)}
\usage{
ssa.btl(f = 10)
}
\arguments{
\item{f}{coarse-graining factor (see page 4 in Chatterjee et al. 2005).}
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\description{
Binomial tau-leap method implementation of the \acronym{SSA} as described by Chatterjee et al. (2005).
}
\references{
Chatterjee et al. (2005)
}
\seealso{
\link{fastgssa-package}, \code{\link{ssa}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
