% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_methods.R
\name{ssa.em}
\alias{ssa.em}
\title{Euler-Maruyama method (EM)

Euler-Maruyama method implementation}
\usage{
ssa.em(h = 0.01, noise_strength = 2)
}
\arguments{
\item{h}{h parameter}

\item{noise_strength}{noise_strength parameter}
}
\value{
an object of to be used by \code{\link{ssa}}.
}
\seealso{
\link{fastgssa-package}, \code{\link{ssa}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
