% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssa_em}
\alias{ssa_em}
\title{Euler-Maruyama method (EM)}
\usage{
ssa_em(h, noise_strength)
}
\arguments{
\item{h}{h parameter}

\item{noise_strength}{noise_strength parameter}
}
\description{
Euler-Maruyama method implementation
}
