% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_etl.R
\name{ssa_etl}
\alias{ssa_etl}
\title{Explicit tau-leap method (ETL)}
\usage{
ssa_etl(tau = 0.3)
}
\arguments{
\item{tau}{the step-size (default 0.3).}
}
\value{
an object of to be used by \code{\link[=ssa]{ssa()}}.
}
\description{
Explicit tau-leap method implementation of the \acronym{SSA} as described by Gillespie (2001).
Note that this method does not attempt to select an appropriate value for tau, nor does it
implement estimated-midpoint technique.
}
\references{
Gillespie D.T. 2001. Approximate accelerated stochastic simulation of chemically reacting systems. J. Chem. Phys. 115:1716-1733. \doi{10.1063/1.1378322}.
}
